% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats.R
\name{Counter}
\alias{Counter}
\title{Counter Stat}
\usage{
Counter(stream = NULL)
}
\description{
A simple stat that counts the number of data points seen so far.
}
\section{Fields}{

\describe{
\item{\code{n}}{Integer count of data points observed.}
}}

\examples{

counter <- Counter()
update(counter, new_x = c(1, 2, 3))
value(counter)

}
