% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base_classes.R
\name{decision}
\alias{decision}
\title{Get the current test decision}
\usage{
decision(test, ...)
}
\arguments{
\item{test}{A Test object}
}
\value{
Character string indicating the current decision:
"Continue" (insufficient evidence), "Reject" (reject the null hypothesis),
or "Accept" (accept the null hypothesis)
}
\description{
Get the current test decision
}
