% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base_classes.R
\name{fetch}
\alias{fetch}
\title{Fetch new data from stream}
\usage{
fetch(stream, ...)
}
\arguments{
\item{stream}{A Stream object}
}
\value{
Transformed records fetched from the stream's current position
}
\description{
Fetch new data from stream
}
