% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base_classes.R
\name{get_length}
\alias{get_length}
\title{Get current length of source}
\usage{
get_length(source, ...)
}
\arguments{
\item{source}{A Source object}
}
\value{
Integer length of data source
}
\description{
Get current length of source
}
