% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base_classes.R
\name{get_n}
\alias{get_n}
\title{Get n records from source}
\usage{
get_n(source, from, n, ...)
}
\arguments{
\item{source}{A Source object}

\item{from}{Starting position, 1-indexed}

\item{n}{Number of records to retrieve}
}
\value{
Vector of records
}
\description{
Get n records from source
}
