% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base_classes.R
\name{get_rest}
\alias{get_rest}
\title{Get remaining records from source}
\usage{
get_rest(source, from, ...)
}
\arguments{
\item{source}{A Source object}

\item{from}{Starting position}
}
\value{
Vector of remaining records
}
\description{
Get remaining records from source
}
