% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base_classes.R
\name{is_stopped}
\alias{is_stopped}
\title{Check if procedure has stopped}
\usage{
is_stopped(stat, ...)
}
\arguments{
\item{stat}{A Stat object}
}
\value{
Logical indicating if stopping condition reached
}
\description{
Check if procedure has stopped
}
