% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base_classes.R
\name{n_obs}
\alias{n_obs}
\title{The number of data points processed so far}
\usage{
n_obs(stat, ...)
}
\arguments{
\item{stat}{A Stat object}
}
\value{
Integer count of data points processed
}
\description{
The number of data points processed so far
}
