% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base_classes.R
\name{reset}
\alias{reset}
\title{Reset object to initial state}
\usage{
reset(object, ...)
}
\arguments{
\item{object}{A Stream or Stat object}
}
\value{
The object (invisibly), with state reset
}
\description{
Reset object to initial state
}
