% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base_classes.R
\name{stopping_time}
\alias{stopping_time}
\title{Get the stopping time of a sequential test}
\usage{
stopping_time(test, ...)
}
\arguments{
\item{test}{A Test object}
}
\value{
Integer giving the number of observations processed when the test
stopped, or \code{NA} if the test has not yet stopped
}
\description{
Get the stopping time of a sequential test
}
