% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base_classes.R
\name{value}
\alias{value}
\title{Get the statistic value(s)}
\usage{
value(stat, n = 1L, ...)
}
\arguments{
\item{stat}{A Stat object}

\item{n}{Number of historical values to retrieve}
}
\value{
The value(s) taken by the statistic
}
\description{
Get the statistic value(s)
}
