% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sources.R
\name{VectorSource}
\alias{VectorSource}
\title{Source data from a vector in an environment.}
\usage{
VectorSource(x, env = parent.frame())
}
\arguments{
\item{x}{The variable name or vector containing the data.}

\item{env}{The environment containing the variable.}
}
\value{
A VectorSource object.
}
\description{
Source data from a vector in an environment.
}
\examples{

x <- c(1, 2, 3, 4, 5)
src <- VectorSource(x)

# Returns c(1, 2, 3)
get_n(src, from = 1, n = 3) 
# Returns c(4, 5)
get_rest(src, from = 4)

}
